Add unit tests for device and data

+++ app/util/database.py
@@ -1,7 +1,8 @@
-import contextlib
-
 from flask import Flask
 from flask_migrate import Migrate
+from random import randint
+from datetime import datetime
+from dateutil.parser import parse
 
 from app.device.meter import Meter
 from app.models import User, Device, Data, Event
@@ -41,12 +42,40 @@ class Database:
 
     @staticmethod
     def cleanDatabase():
+        # If the database is not the test one, avoid using this
+        if str(db.engine.url).find('app-test.sqlite') == -1:
+            raise RuntimeError()
+
         db.session.query(Device).delete()
         db.session.query(Data).delete()
         db.session.query(Event).delete()
         db.session.query(User).delete()
         db.session.commit()
 
+    @staticmethod
+    def addTestingData(device,
+                       start_date='2015-01-01',
+                       end_date='2015-01-30',
+                       data_count=1000,
+                       min_value=1,
+                       max_value=100,
+                       min_change=-5,
+                       max_change=5):
+        start_time = parse(start_date).timestamp()
+        end_time = parse(end_date).timestamp()
+        data_value = randint(min_value, max_value)
+
+        for timestamp in range(int(start_time), int(end_time), int((end_time - start_time) / data_count)):
+            data = Data(
+                time=datetime.fromtimestamp(timestamp),
+                value=data_value,
+                device_id=device.id
+            )
+            data_value = min(max(data_value + randint(min_change, max_change), min_value), max_value)
+            db.session.add(data)
+
+        db.session.commit()
+        db.session.flush()
 
     @staticmethod
     def initiateDatabase(app):
